Create README.md
It's a simple timetable app with several features. You just need to fill your timetable in the 'Edit Timetable' tab and there you go. Now you need to click a button on the home page tab to see your current day's timetable. On clicking the top left corner, the navigation drawer opens which contains all the week days from Monday to Saturday, you can click on any one of them to know that particular day's timetable. The interesting feature of this app is that once you have created your timetable you can share it with anyone and then they doesn't have to create their own, they just need to import it into the app from the folder in which they have got the timetable. You can also create the backup of your timetable, it will be stored in a folder named 'TimeTableApp' in the internal storage of your device, so that if any how your app is accidentally uninstalled, just reinstall it and you can directly import your timetable from the 'TimeTableApp' folder into the app. And you can also change the background of your home page.